ENH: Added changes required for integrating Metric examples with Midas.
To integrate the ValidateDice and ValidateJaccard example applications with
Midas I needed to change the applications to output to a file, rather than
outputting to standard out (standard out makes it difficult to grab the
output after executing on condor).  This change also required a change to the
generateclp xml files for these applications.

I had to create a bmm file, a BatchMake script bms file, and a Midas
application description xml file for each of the example applications, so
that they could be run from the Midas-BatchMake plugin.

Finally, I had to modify the CMakeLists.txt files to copy over these
three required Midas-Integration files to the build location.